WIP: feature/vistoria-com-anexos
