+++ src/app.module.ts
@@ -10,15 +10,14 @@ import { Prisma2Module } from './prisma2/prisma2.module';
 import { VitalidadeModule } from './vitalidade/vitalidade.module';
 import { ConfigModule } from '@nestjs/config';
 import { CadastrosModule } from './cadastros/cadastros.module';
-import { IntegracoesModule } from './integracoes/integracoes.module';
 import { VistoriasModule } from './vistorias/vistorias.module';
+import { IntegracoesModule } from './integracoes/integracoes.module';
 import { ProspeccoesModule } from './prospeccoes/prospeccoes.module';
 import { SalvaguardaModule } from './salvaguarda/salvaguarda.module';
 import { DevtoolsModule } from '@nestjs/devtools-integration';
 import { AuditoriasModule } from './auditorias/auditorias.module';
 import { BuscasModule } from './buscas/buscas.module';
 import { RelacionamentosModule } from './relacionamentos/relacionamentos.module';
-import { Vistoria2Module } from './vistoria2/vistoria2.module';
 
 @Global()
 @Module({
@@ -40,7 +39,6 @@ import { Vistoria2Module } from './vistoria2/vistoria2.module';
     AuditoriasModule,
     BuscasModule,
     RelacionamentosModule,
-    Vistoria2Module,
   ],
   controllers: [],
   providers: [
