+++ src/vistorias/vistorias.module.ts
@@ -1,12 +1,10 @@
 import { Module } from '@nestjs/common';
 import { VistoriasService } from './vistorias.service';
 import { VistoriasController } from './vistorias.controller';
-
-import { SalvaguardaModule } from 'src/salvaguarda/salvaguarda.module';
+import { SalvaguardaService } from 'src/salvaguarda/salvaguarda.service';
 
 @Module({
-  imports: [SalvaguardaModule],
   controllers: [VistoriasController],
-  providers: [VistoriasService],
+  providers: [VistoriasService, SalvaguardaService],
 })
 export class VistoriasModule {}
